+++ Vigil/build.rs
@@ -1,8 +1,8 @@
 fn main() {
     let mut res = winres::WindowsResource::new();
-    res.set("ProductName", "TITAN Vigil");
-    res.set("FileDescription", "TITAN Vigil");
-    res.set("CompanyName", "TITAN");
+    res.set("ProductName", "RYFTENIUS Vigil");
+    res.set("FileDescription", "RYFTENIUS Vigil");
+    res.set("CompanyName", "RYFTENIUS");
     res.set_manifest_file("windows/vigil.manifest");
     res.compile().unwrap();
 }